# led.c: update blink timings for better comprehension