--- a/src/touch.h
+++ b/src/touch.h
@@ -32,7 +32,7 @@
 
 #define QTOUCH_TOUCH_TIMEOUT            3000// msec
 #define QTOUCH_TOUCH_TIMEOUT_HARD       30000
-#define QTOUCH_TOUCH_BLINK_ON           3000
+#define QTOUCH_TOUCH_BLINK_ON           4000
 #define QTOUCH_TOUCH_BLINK_OFF          300
 #define QTOUCH_TOUCH_THRESH_HW_V1_2     18
 #define QTOUCH_TOUCH_THRESH             15 // hw version 1 / unknown