- updated function naming for better dev understanding
- blink off increased from 100ms to 200ms
- blink on remains at 100ms
- increased the overall pattern period from 3 sec to 4 sec